+++ resources/frontend/admin/src/components/form/mixins/FieldMixin.vue
@@ -28,7 +28,7 @@
       },
       validationRule: {
         type: [String, Object, Array],
-        default: () => ([]),
+        default: '',
       },
       attributes: {
         type: [Array, Object],